developer.wordpress.org: don't make parameters without explicit defaults appear as though they are defaulted to null
In the output of method signatures, it is misleading for the lack of a default value for a parameter to be presented as if the parameter was assigned a default value of null. At first glance it gives the impression the parameter is optional (it isn't) or that if a null value is explicitly provided the function will behave in its default manner (it likely won't).