Theme Directory: Remove bundled trac class.
There is an IXR class on dotorg that is the same as the one in core, except
that it was modified to handle SSL connections. The Trac class was originally
written with this dotorg IXR class in mind. Here, we're switching to that class
and add support for WP_IXR_HTTP_Client, which is what WordPress uses when it
uses the IXR class, so we can connect to Trac installs that are SSL.
Soon, this Trac library will be placed in a common includes directory; for now,
we update it to use what is already on dotorg elsewhere.