WP.org Themes: Make sure uploaded theme files are always deleted.
We were already at a state where files were always deleted, independent from
the outcome of automated checks. Hocking it to PHP shutdown let's us clean up
in one place, and we don't have to worry about remembering to add it in when
other checks are added or existing checks modified.

See #745.