WP.org Themes: Remove approval workflow.
Theme Review Admins do a final check on Trac and set themes live from there.
For the off chance that a ticket needs manual adjustment, version statuses can
be changed in the edit screen of a repopackage item.

This also removes Trac syncing, in order for it to be added as a script on the
dotorg side.

See https://wordpress.slack.com/archives/meta/p1423071510000491
See #745.