WordCamp JSON API: Create whitelisted endpoint array from scratch.
This essentially reverts [wordcamp1972], so that we are creating a new array and populating it with whitelisted items, instead of traversing the original array and removing blacklisted items.

The previous approach was taken because at the time it appeared we would need to access endpoint variables that were not available globally -- such as `$wp_json_post_meta` -- but it turns out that we don't, and this approach is much simpler. In order to add new endpoints like `posts/types`, which have a different array structure than the currently supported ones, it would have been necessary to significantly increase the complexity of the traversal logic. With this approach, adding new endpoints is straightforward, regardless of structural inconsistencies.