WordCamp Organizer Nags: Migrate admin notices to a dashboard widget.
Admin notices should only be used for relatively urgent messages, not day-to-day tasks, and they should only appear on screens that are relevant. Global, non-dismissable notices with mundane content are counter-productive in the long run, because they train people to ignore all notices.

A dashboard widget is a much more appropriate and sustainable medium for this type of communication.