WordCamp Post Type: Pull organizer username from `$_POST` instead of post meta.
`$_POST` will always contain the latest value, but post meta could be outdated if the value is changed during the same request that the post transitions from `draft` to `pending`. This often resulted in the organizer not being added because the field was empty, and then it was filled it at the same time that the post transitioned.