WordCamp Post Type: There's no reason to go through five filters to define a constant.
Clean up a lot of the old over-engineered code, remove
unused actions and filters, actions that are wrappers for
other core actions and other things.

If your code relies on any of these actions and filters,
feel free to bring it back, or better yet change the
action/filter name to the one provided by core.