Plugin Directory: Allow Committers and Reviewers to `edit_others_posts`.
WordPress seems to require users to have that capability globally for a post
type, in order to make changes to others posts, even if it's only certain
others posts. 

This switches to using `plugin_review` and `plugin_approve` capabilities to
control Reviewers and Committers access to certain plugins. It also improves
the logic for displaying links to the various post status views in the plugins
list table.

See #1570.