Plugin Directory: Proof of concept React components.
This is pretty much the culmination of me playing around with React for a week,
learning along the way. It's a best of from tutorials, Human Made's
Feelingrestful Theme, and various other inspirations I got along the way.

I'm committing this not with the intention of cementing a way to build a JS
client for the Plugin Directory but much rather open this approach up for tips
and feedback on how to proceed. I'd like to find out if React is even working
for our needs, learn how to get the most out of it for our purposes, and learn
how best to implement it for this project.

To see the React client in action and make changes to it, uncomment the
`wp_enqueue_script()` line in `functions.php` that enqueues the client JS file.
Then run `npm install` and `grunt watch` from the command line in the theme's
root directory. Please see `README.md` for further details.

See #1719.