Plugin Directory: Readme parsing: When parsing the FAQ section, pull out any properly formatted (and some not-so-well-formatted) Question & Answer pairs to be displayed in a formatted fashion.
The primary method we support for FAQ entries is where the sub headings are the question, with the content following as the answer. Both `=` and `#` are supported for this: 
{{{
=== FAQ ===
== Does the plugin do X? ==
Yes!
## What about Y?
Nah, Sorry!
}}}

As plugins have a wide range of contents at present, we also support plugins which have used bolded headingslike the following, note, that it must start and end the line with two asterisks.
This format isn't suggested for future usage, and is only included for the thousand or more common plugins which have used it.
{{{
=== FAQ ===
** Can I contribute to the plugin? **
Yes! Patches Welcome!
}}}

Finally, there's the plugin readme's who don't confirm to any of the above standards, or have included a paragraph of data prior to their Q&A's, for those plugins the freeform content will be prepended to the FAQ section and they won't get a properly formatted accordian-style interface.

Props obenland for the initial patch, a lot of subtle changes have been made to better handle the various readme's out there (and probably more tweaks are needed).

See #1810