Events: Skip IP searches when locations are passed
Otherwise, users could be confused and frustrated when they search for a city and end up receiving a completely different city. Behind the scenes, the `location_name` search failed, and the IP search succeeded, but that's not apparent to the user.

It's safe to assume that if someone is searching for a city, they don't want to use a location based on their IP.