WordCamp Central: Add year to camp dates on the schedule
Changes the Schedule page template to emulate the organization on the
Past Camps page template, by grouping camps under a year header and
including the year in each camp's date.

This also adds a param to `WordCamp_Central_Theme::the_wordcamp_date` to
include the year in the camp date output. The default is to not include
it, for backcompat.

Props frq
Fixes #2065