WordCamp Post Types: Consolidate v2 REST functionality, parity w/ v1
This moves all the actions/filters related to v2 REST API endpoints
to a separate file and adds the (hopefully) last bit of response data
that was included in v1, but was missing in v2 (speaker avatars).

For the speaker avatar data, I decided to emulate how Core presents
the avatar data for post authors in API responses. So there is a top-
level `avatar_urls` field that is an object containing multiple
Gravatar URLs for different image sizes.