WordCamp Central: Restore the About map cache expiration to 1 week.
About 15 seconds after deploying r6665, I remembered why the original expiration was 1 week: because there are multiple maps using this function, and the one on the About page is not visited as often, and doesn't need to be updated as often. It's better to make sure it loads quickly than to make sure it's always up to date.

The map on the Schedule page (and any other maps that are added in the future) will still use a 1 day expiration.

See #3413