Gutenberg: Override `wp.domReady` with a version which doesn't contain a race-condition, this fixes the whitescreen on frontendberg on first load.
`wp.domReady()` is incorrectly regisering callbacks (such as the Gutenberg init function) to be fired on `DOMContentLoaded` AFTER that event has already fired, causing the callbacks to never run.
This replaces it with a version which does work, calling the callback immediately when `document.readyState` is at `interactive` (which is what fires `DOMContentLoaded`) or `complete` (when styles/images/etc are loaded, which the Gutenberg `wp.domReady` does).

See #3703, #3735.