WordCamp Base: Use gallery term ID to avoid slug mismatch fatal error.
Previously, the code assumed that if a term existed with the name `Gallery`, then it's slug would also be `gallery`. If it didn't match, `get_term_link()` would return a `WP_Error`, because it won't search by name like `in_category()` will. Trying to convert that `WP_Error` to a string to `echo` resulted in the fatal error.