WordCamp Tickets: Change the Attendees page checkbox to yes/no radios
This changes the input for the ticket registration question, "Do you want to be
listed on the public Attendees page?" from a checkbox to a Yes/No pair of
radio inputs. The radios do not have a default value, but it is a required
field in the form, so the attendee must make a choice before they can proceed
with buying a ticket. This change is due to concerns raised in this discussion:
https://github.com/Automattic/camptix/issues/184