Time Shortcode: properly override the `$shortcode_tags` array when running the time shortcode in comments early.
Calling `unset()` on the global array wasn't emptying it out, but rather unsetting it in the scope of the function, causing `do_shortcode()` to run with the original shortcode arrays.

This fixes a case where `[code]` shortcodes were being run before the plugin responsible was able to hook in (early) to handle the shortcodes.

H/t tobifjellner for noticing the problem.
See https://wordpress.slack.com/archives/C02QB8GMM/p1543802741017900