WordCamp Payment: Send reminder mails to organizer about pending invoices.
Sometimes a sponsor invoice can be in pending state for too long, in which case either organizers need to cancel the invoice or they need to nudge the sponsors to pay them. This patch sends a reminder mail to organizer about an invoice that has been pending for too long. After 60 days of being sent, or after 60 days of WordCamp start date, we will send a mail to WordCamp support notfying them about the invoice.