WordCamp Organizer Reminders: Use empty array as default sent email IDs.
The previous code attempted to do this by casting the result to an array, but that accidentally created an array where the first value was an empty string -- i.e., `array( 0 => '' )` -- because `get_post_meta()` returns an empty string when `$single` is `false`.

In order to correctly end up with an empty array, the code needs to explicitly check that the returned value isn't already an array.