Session Block: Pass props to fix `attributes undefined` bug.
r8506 added support to directly changing `featured_image_width` attribute and directly uses this attribute to render image width. This caused a bbug for when `attribute` is not passed in props preventing rendering of block inside editor. This patch fixes that by passing props. Another fix is required to add sanity check in FeaturedImage component.