WC Report: Use `return` b/c this action may be hooked to `admin_init`.
In r7902, we added a `wp_die` statement if nonce is not verified for Meetup Events report. However, when opening WordCamp Report, this method gets hooked to `admin_init`, and cause all reports that are loaded after it, like Meetup Details report to break.

Replacing `wp_die` with `return` makes the behavior consistent with other reports and fix this bug.