Theme Diretory: When creating a Trac ticket for a theme, link to the Theme in the description.
Previously Theme Trac tickets did not include any form of link back to the theme directory, which can make debugging harder.

For new themes, before they're live, this link won't work, but is included for when the theme is made live and as a preview of the themes final location.