Official WordPress Events: Tweak the logic around token refreshing.
The cron tasks were failing sporadically as the token wasn't valid - it looks like the token "looked valid" and expired, but looks like the refresh token was also expired.

This feels lik a hammer, forcing a new token to be fetched when both the token and refresh token are expired, but it worked.