Official WordPress Events: Sync WordCamps in "pre-planning"
This is a workaround for the Events Widget. WordCamps that have been moved from scheduled back to pre-planning (due to COVID-19) need to be marked as un-scheduled in the wporg_events table, so that the API doesn't return them as still running. This change syncs "In Pre-Planning" events as a new status, `pre-planning`.