Login: Allow users to resend the signup confirmation email.
To trigger this, either the user needs to click the resend link on the post-signup profile-info page, or attempt to sign up again using the same username or email.

The emails are rate limited depending on  the age of the account, and also based on some WordPress.org anti-spam measures.

See #5278.