Plugin Directory: Add a check for one top-level block.
Multiple blocks are allowed, but there should be only one "main" block. This checks the block.json file for each block (if exists), and flags if there is more than 1 block without a parent property. This also fixes the `check_for_blocks` check to warn if more than 15 blocks are found.

See #5303