Make: Re-enable r10309 with additional fixes for o2.
o2 outputs a hidden JSON "post fragment" when `the_content()` is called so that the post ends up in the JS render.
Due to the global `$post` not being set previously, the fragment being output was ignored as it didn't contain any valid data, after the change.. it did.

This change tells o2 that the fragment for the welcome post is not to be used, by returning an empty array.

This should fix the Notices, and not result in the welcome box being shown inline on the rendered o2 page.