Theme Preview: Starter Content: Making `is_customize_preview()` return truthful instead of introducing a new conditional like `is_wpthemes()` did have some side effects.
Some themes assume the customizer is loaded based on `is_customize_preview()` without checking to see if the classes exist, or waiting until an approriate hook to fire, resulting in a fatal error when the class isn't available.

Work around that, by aliasing a bunch of other WP_Customize_* classes. This probably needs a re-think, but this works fairly well for now.