Theme Directory: Tell WordPress to use `index.php` as the template for archive & singular views.
This is required as the theme is now marked as a child theme of `wporg`, and WordPress will fall back to it's templates if it doesn't exist in the child theme.
The Theme Directory being a special theme, most views simply aren't defined as it's intended that the index.php template would have been the fallback for them.