Main theme: Change semantics of some containers on the Download page
The Download page has info about the mobile apps and hosting for WordPress.
These are in `section` container elements, which might be the reason that Google
is using the content of the hosting container for its search result snippet,
instead of the meta description tag. By changing these containers to `aside`
elements, hopefully Google will get the message that they don't contain the
most pertinent information for that page.

props dd32, coreymckrill