Translate: Explicitely exclude `patterns` from `make-core-pot` more firmly by ignoring the `wp-content` directory entirely for these admin contexts.
It appears that the command does not fully respect the `exclude` parameter if it believes there are strings that should be included. By excluding a higher-level directory, it never see's the potential included `patterns` folder.

Ensures that strings from Twenty Twenty-Two are not included in POT files for wp-admin.

Previously: [11809].
See https://github.com/wp-cli/i18n-command/pull/312#pullrequestreview-959463247