Slack: Announce: Properly handle private channels.
Older slack private channels send their channel_name as 'privategroup', but newer slack channels send their actual name.
To resolve this, we'll check the Slack API to verify the channels metadata.

The `/here` handler was getting caught on this fine detail, and treating some (but not all!) private groups as public channels.
This is what caused a private channel to broadcast to a public channel, and an incorrect fix was tried via r11875.

Reverts [11875].