GP Profiles: Remove unnecessary checks to include bulk reject/fuzzy.
The first argument to `GP::$permission->current_user_can()` needs to be `edit`, not `reject` / `fuzzy`. That bug was causing `$activities` to be empty on reject/fuzzy reviews.

`add_bulk_translation_activity` is only fired when the current user has permission to approve strings in the translation set, though, so checking for `edit` here is redundant in practice.