Photo Directory, Uploads: Add JS pre-upload validation to check if submission has same file name as another pending submission from the contributor.
The check is more of a means to immediately alert contributor to a likely duplicate file submission without having to wait upon submission for the actual duplication check. An identical server-side check for this is unnecessary as the file will be checked for actual duplication.

Props roytanck, coffee2code.
See #6537.