Support Forums: Plugin/Theme/Review "forums": Make `bbp_get_forum_id()` work in the plugin/theme/review "forums".
These forums are not forums, but rather, are views of a forum. As a result of this, bbPress doesn't return a forum object even through it's seen as a single forum.

This caused the Block Editor to be enabled for these views due to a misconfiguration of the Blocks integration, where in the event it couldn't find a forum for the current page, defaulted to being enabled.

This code aims to remove similar faults in the future, by allowing `bbp_get_forum_id()` to return the true forum_id for these forums, while at the same time being a view.

This should not be removed without verifying that [12348] works as expected.

See #6608.