Plugin Directory: Add ChatGPT to protected terms
Sadly this is the new hot-topic to make plugins for, by people who are not as familiar with the restrictions with names.
In order to not have to explain the same thing and keep pending plugins, this is going to just stop 'em up front.
I put in a secondary spelling to get ahead on that one.

Fixes #6753